fix: followingCount/followerCount 미반영 문제 해결 (#187)

+++ src/main/java/com/ktb3/devths/user/dto/response/UserSignupResponse.java
@@ -22,6 +22,6 @@ public record UserSignupResponse(
 	public record ProfileImage(Long id, String url) {
 	}
 
-	public record UserStats(int followerCount, int followingCount) {
+	public record UserStats(long followerCount, long followingCount) {
 	}
 }
